Bump up smtp_accept_max_per_host
Exim can have a limit on the number of onnections per host.  It was set
fairly low as a defensive measure, but it turns out Airmail.app when
doing auto-discovery of settings will connect in paralell and
effectively ddos itself out of business.  4 seems to be a safe
compromise here.