Deliver Spam to ~/Maildir/.Spam rather than ~/Maildir/.INBOX.spam
In switching to Apple Mail on the desktop the wisdom of pretending the
junk mail folder is a sub-folder of the INBOX folder comes into
question.  Unread mail in the .INBOX.spam box counts towards the unread
count of the INBOX in total, so makes INBOX zero further away and more
of a morass than it needs to be.

Solution: filter into =.Spam instead, and then add that to mutt's list
of monitored boxes for when we want to find it quickly in mutt.