emacs: add org-protocol, capture templates
org-protocol allows for specific org-mode functions to be invoked via
emacsclient with a uri structure.  For example a browser extension or
bookmarklet recording a url might call
org-protocol:/capture:/$template/$url/$title

We also add some capture templates, that the chrome plugin expects to
be there for capturing notes and/or links.

As the bookmarks org file uses ox-html to export a pretty/searchable
page, fix it by specifying org-export-async-debug nil.