tmux/zsh/nvim: Stomp over $TERM less
nvim really relies on the TERM setting to reflect that the terminal it's
on supports truecolor/RGB colors.  our hackaround for oh-my-zsh's title
setting was leading to TERM getting squashed down to 'xterm', which
doesn't have the RGB features.

The new hack is to have tmux advertise itself as xterm-256color to zsh,
as then the terminal helper in oh-my-zsh acts correctly (from
screen-256color it's not sending both window and tab name)
https://github.com/ohmyzsh/ohmyzsh/blob/master/lib/termsupport.zsh#L19

The effect of this is nvim can see the terminal it's in has RGB support,
oh-my-zsh sets titles in a way tmux understands.  Something will
probably break in the gap between screen-256color and xterm-256color,
but that's tomorrow's headache.