tmux/kitty/nvim: tomorrow's headache today
In 912199 we had tmux start advertising itself as xterm-256color to make
oh-my-zsh work, and noted it might cause a future headache.

Well the future came quickly.  If we use kitty, then launch tmux, and
then nvim, the colors are screwed up.  TERM of screen-256color fixes that,
but stops the terminal title stuff stop working.

As we're somewhat partial to the terminal title stuff, we might be
better off just dropping oh-my-zsh, and then reimplement the bits we
really use from it in a simple _zshrc.