zshrc: adapt to NixOS
Two main changes:
	Stops stomping the PATH
	Drops and reimplements settings from oh-my-zsh

The PATH changes were needed as we were stomping all over `nix shell`'s
dynamic PATH setting, so the flake experience was poor.

The other more sizable things is throwing away oh-my-zsh.  Outside of a
pretty prompt I wan't getting much use from it, and couldn't be bothered
to reinstall on every machine rebuild.

We'll probably configure starship later for a prettier prompt, but gate
that behind a has check, so on machines we don't do a lot of shell work
on, we shouldn't miss it.