tmux: pass current path over all OS
The original hack with testing uname==Darwin was that the macbook had a
newer tmux, and so we were trying to gate the use of `-c #{pane_current_path}`
to 'new' hosts, so not to break the aged tmux installs on some debian boxes.

We've upgraded past those now, so we can drop the hacks.