nvim: NixOS rust-analyzer - make mason append to PATH
The default for mason is to prepend to PATH, which makes the binaries that mason
installs take priority over any you might have installed elsewhere.

That's normally what you want, but on NixOS it breaks down because the
mostly-statically-linked rust-analyzer expects /lib64/ld-linux-x86-64.so.2 to
be present.

  $ file -L ~/.local/share/nvim/mason/bin/rust-analyzer
  /home/richardc/.local/share/nvim/mason/bin/rust-analyzer: ELF 64-bit LSB pie executable, x86-64, version 1 (SYSV), dynamically linked, interpreter /lib64/ld-linux-x86-64.so.2, for GNU/Linux 3.2.0, BuildID[sha1]=657974cba1e25b7c03f5b14cd00bc426ffc05dee, with debug_info, not stripped

But as this is NixOS, /lib64 isn't there, and trying to run that binary
gets a fun error as the 'interpreter' (ld-linux) isn't there:

  $ ~/.local/share/nvim/mason/bin/rust-analyzer
  zsh: no such file or directory: /home/richardc/.local/share/nvim/mason/bin/rust-analyzer

We do have a working (but old) version available via `rustup component add rust-analyzer`

  $ file -L $(which rust-analyzer)
  /run/current-system/sw/bin/rust-analyzer: ELF 64-bit LSB executable, x86-64, version 1 (SYSV), dynamically linked, interpreter /nix/store/1zy01hjzwvvia6h9dq5xar88v77fgh9x-glibc-2.38-44/lib/ld-linux-x86-64.so.2, for GNU/Linux 3.10.0, not stripped

So we can make use of that version if don't prepend to PATH.

A better fix might be to detect NixOS and then simply not use mason
there.  There may also be mileage using nix-ld/nix-alien to paper
the gap.